build: Generate ChangeLog and add it to the tarball
* .gitattributes: New.
* CMakeLists.txt: Include G10AddGenChangeLogTarget. Add "gen-ChangeLog"
target. Make "dist" target depend on "gen-ChangeLog" target.
* cmake/modules/G10AddGenChangeLogTarget.cmake,
cmake/modules/g10_generate_ChangeLog.cmake.in: New.
--

This adds a custom "gen-ChangeLog" target that can be used to generate
a ChangeLog from the commit log messages. The ChangeLog starts at the
import of the sources from the KDE repo in the gpgme repo because the
commit log messages in the KDE repo don't use the proper formatting.

GnuPG-bug-id: 7262