Modernize interface of QuickJob::startCreate and ::startAddSubkey
* src/quickjob.h (QuickJob::startCreate, QuickJob::startAddSubkey): Add
overload. Deprecate old overload and make it non-virtual.
* src/quickjob_p.h (class QuickJobPrivate): Add pure virtual methods
startCreate and startAddSubkey.
* src/quickjob.cpp (QuickJob::startCreate, QuickJob::startAddSubkey):
New.
* src/qgpgmequickjob.h, src/qgpgmequickjob.cpp
(QGpgMEQuickJob::startCreate, QGpgMEQuickJob::startAddSubkey): Remove.
* src/qgpgmequickjob.cpp (QGpgMEQuickJobPrivate::startCreate,
QGpgMEQuickJobPrivate::startAddSubkey): New.
(createWorker): Change type of algo argument and of flags argument.
Remove key argument. Call the new Context::createKey overload.
(addSubkeyWorker): Change type of algo argument and of flags argument.
Call the new Context::createSubkey overload.
--

This replaces the usage of the deprecated overloads of createKey and
createSubkey and aligns the interface of the functions to the changes
in GpgME++.